Add utils, rename static variables upper case
